Big changes in architecture

+++ InterferometricSignalsProcessing/FilterTuning/GradientTuner.h
@@ -8,22 +8,22 @@ namespace FilterTuning
 	class GradientTuner
 	{
 	public:
-		GradientTuner(double **inputSignals_, int signalSize_, int signalsCount_, int iterationsCount_,
+		GradientTuner(float **inputSignals_, int signalSize_, int signalsCount_, int iterationsCount_,
 			ExtendedKalmanFilterIS1DState currentState_, ExtendedKalmanFilterIS1DState step_);
 		~GradientTuner();
 
-		void changeSignals(double **inputSignals_, int signalsCount_);
+		void changeSignals(float **inputSignals_, int signalsCount_);
 		ExtendedKalmanFilterIS1DState tune();
 		void makeStep();
 	private:
-		double **inputSignals;
+		float **inputSignals;
 		int signalSize;
 		int signalsCount;
 		int iterationsCount;
 		ExtendedKalmanFilterIS1DState currentState;
 		ExtendedKalmanFilterIS1DState step;
 
-		int gradSign(double s, double interval = 1.);
+		int gradSign(float s, float interval = 1.);
 	};
 }
