+++ InterferometricSignalsProcessing/FilterTuning/TotalSearchTuner.h
@@ -4,26 +4,33 @@
 #include <random>
 
 #include "../Filters/ExtendedKalmanFilterIS1D.h"
+#include "../DataModel/SignalAnalysis.h"
 
 namespace FilterTuning
 {
 	class TotalSearchTuner
 	{
 	public:
-		TotalSearchTuner(double **inputSignals_, int signalSize_, int signalsCount_, int filtersCount_,
+		TotalSearchTuner(float **inputSignals_, int signalSize_, int signalsCount_, int filtersCount_,
 			std::default_random_engine &gen_, ExtendedKalmanFilterIS1DState min_, ExtendedKalmanFilterIS1DState max_);
 		~TotalSearchTuner();
 
 		void createStates();
-		void changeSignals(double **inputSignals_, int signalsCount_);
+		void changeSignals(float **inputSignals_, int signalsCount_);
 
 		ExtendedKalmanFilterIS1DState tune();
 	private:
 		ExtendedKalmanFilterIS1DState createRandomState();
-		double getRandom(double min, double max);
+		float getRandom(float min, float max);
 
-		double **inputSignals;
+		//todo
+		float **inputSignals;
 		ExtendedKalmanFilterIS1DState *filterStates;
+		
+		//new
+		//std::vector<dmod::signal1d> inputSignals;
+		//std::vector<ExtendedKalmanFilterIS1DState> filterStates;
+		//
 		int signalsCount;
 		int signalSize;
 		int filtersCount;
